% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_classify.R
\name{mod_classify_ui}
\alias{mod_classify_ui}
\alias{mod_classify_server}
\title{mod_classify_ui and mod_classify_server}
\usage{
mod_classify_ui(id)

mod_classify_server(
  input,
  output,
  session,
  r,
  img,
  cell_seg,
  ph_norm,
  classify
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
