% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_classify_loop.R
\name{mod_classify_loop_ui}
\alias{mod_classify_loop_ui}
\alias{mod_classify_loop_server}
\title{mod_classify_loop_ui and mod_classify_loop_server}
\usage{
mod_classify_loop_ui(id)

mod_classify_loop_server(input, output, session, r, classify)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
