% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cmodel_dir.R
\name{mod_cmodel_dir_ui}
\alias{mod_cmodel_dir_ui}
\alias{mod_cmodel_dir_server}
\title{mod_cmodel_dir_ui and mod_cmodel_dir_server}
\usage{
mod_cmodel_dir_ui(id)

mod_cmodel_dir_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
