% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_create_model.R
\name{mod_create_model_ui}
\alias{mod_create_model_ui}
\alias{mod_create_model_server}
\title{mod_create_model_ui and mod_create_model_server}
\usage{
mod_create_model_ui(id)

mod_create_model_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
