% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_ch.R
\name{mod_display_ch_ui}
\alias{mod_display_ch_ui}
\alias{mod_display_ch_server}
\title{mod_display_ch_ui and mod_display_ch_server}
\usage{
mod_display_ch_ui(id)

mod_display_ch_server(input, output, session, img, n, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
