% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dv_input.R
\name{mod_dv_input_ui}
\alias{mod_dv_input_ui}
\alias{mod_dv_input_server}
\title{mod_dv_input_ui and mod_dv_input_server}
\usage{
mod_dv_input_ui(id)

mod_dv_input_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
