% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_img_dir.R
\name{mod_img_dir_ui}
\alias{mod_img_dir_ui}
\alias{mod_img_dir_server}
\title{mod_img_dir_ui and mod_img_dir_server}
\usage{
mod_img_dir_ui(id)

mod_img_dir_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
