% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_img.R
\name{mod_load_img_ui}
\alias{mod_load_img_ui}
\alias{mod_load_img_server}
\title{mod_load_img_ui and mod_load_img_server}
\usage{
mod_load_img_ui(id)

mod_load_img_server(input, output, session, r, ix)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
