% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_testfiles.R
\name{mod_load_testfiles_ui}
\alias{mod_load_testfiles_ui}
\alias{mod_load_testfiles_server}
\title{mod_load_testfiles_ui and mod_load_testfiles_server}
\usage{
mod_load_testfiles_ui(id)

mod_load_testfiles_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
