% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_n_segment.R
\name{mod_n_segment_ui}
\alias{mod_n_segment_ui}
\alias{mod_n_segment_server}
\title{mod_n_segment_ui and mod_n_segment_server}
\usage{
mod_n_segment_ui(id)

mod_n_segment_server(input, output, session, params, nuc_norm)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
