% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nuc_params.R
\name{mod_nuc_params_ui}
\alias{mod_nuc_params_ui}
\alias{mod_nuc_params_server}
\title{mod_nuc_params_ui and mod_nuc_params_server}
\usage{
mod_nuc_params_ui(id)

mod_nuc_params_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
