% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ph_segment.R
\name{mod_ph_segment_ui}
\alias{mod_ph_segment_ui}
\alias{mod_ph_segment_server}
\title{mod_ph_segment_ui and mod_ph_segment_server}
\usage{
mod_ph_segment_ui(id)

mod_ph_segment_server(input, output, session, r, params, ph_norm, nseg)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
