% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_ch.R
\name{mod_select_ch_ui}
\alias{mod_select_ch_ui}
\alias{mod_select_ch_server}
\title{mod_select_ch_ui and mod_select_ch_server}
\usage{
mod_select_ch_ui(id)

mod_select_ch_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
